+++ src/components/pages/common/index.tsx
@@ -3,11 +3,13 @@ import QRNotFound from './qr-not-found'
 import QRNoConnection from './qr-no-connection'
 import QRIncorrectParameter from './qr-incorrect-parameter'
 import ERC20TokenPreview from './erc20-token-preview'
+import ClaimingFinishedButton from './claiming-finished-button'
 
 export {
   QRNotMapped,
   QRNotFound,
   QRNoConnection,
   QRIncorrectParameter,
-  ERC20TokenPreview
+  ERC20TokenPreview,
+  ClaimingFinishedButton
 }