+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-# Ledger Claim app
+# Linkdrop Claim app
+
+## 2.2.1-rc.1
+- Alchemy API added
+- ERC20 claim flow
+- Updates for error handling
+- Support for claiming finished params
 
 ## 2.2.0